+++ Libraries/shared-code/crossbow-core/main/src/com/crossbow/volley/toolbox/CrossbowComponents.java
@@ -1,6 +1,8 @@
 package com.crossbow.volley.toolbox;
 
 
+import com.android.volley.Cache;
+import com.android.volley.Network;
 import com.android.volley.RequestQueue;
 import com.crossbow.volley.CrossbowImageCache;
 import com.crossbow.volley.FileQueue;
@@ -12,4 +14,8 @@ public interface CrossbowComponents {
     CrossbowImageCache provideImageCache();
 
     NetworkImageLoader provideImageLoader();
+
+    Cache provideCache();
+
+    Network provideNetwork();
 }